+++ example/test_app/views.py
@@ -1,5 +1,10 @@
+from django.http import HttpResponseRedirect
 from django.shortcuts import render
 from django.contrib.auth.decorators import login_required
+
+
+def public(request):  # pragma: no cover
+    return render(request, "public.html", {})
 
 
 @login_required()
@@ -10,3 +15,17 @@ def home(request): # pragma: no cover
 @login_required()
 def registered(request): # pragma: no cover
     return render(request,"home.html",{"registered":True})
+
+def manage(request):
+    if not request.user.is_authenticated:
+        return HttpResponseRedirect("../login/")
+    return render(request,'rest/manage.html',{})
+
+def login(request):
+    if request.user.is_authenticated:
+        return HttpResponseRedirect("../manage/")
+    return render(request,'rest/login.html',{})
+
+
+def choose(request):
+    return render(request,'choose.html',{})