+++ passkeys/backend.py
@@ -1,4 +1,6 @@
 from django.contrib.auth.backends import ModelBackend
+from django.core.exceptions import ImproperlyConfigured
+
 from .FIDO2 import auth_complete
 
 class PasskeyModelBackend(ModelBackend):